moved network and database logic to repository
