+++ app/src/androidTest/java/com/sourabh/xplayer/ExampleInstrumentedTest.kt
@@ -1,7 +1,7 @@
 package com.sourabh.xplayer
 
-import android.support.test.InstrumentationRegistry
-import android.support.test.runner.AndroidJUnit4
+import androidx.test.InstrumentationRegistry
+import androidx.test.runner.AndroidJUnit4
 
 import org.junit.Test
 import org.junit.runner.RunWith
