+++ build.gradle
@@ -3,10 +3,13 @@
 buildscript {
     ext.kotlin_version = '1.3.31'
     ext.kotlin_version = '1.3.11'
+    ext.room_version = '2.1.0'
+    ext.lifecycle_version = '2.0.0'
+
     repositories {
         google()
         jcenter()
-        
+        mavenCentral()
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:3.4.1'
@@ -20,7 +23,7 @@ allprojects {
     repositories {
         google()
         jcenter()
-        
+        mavenCentral()
     }
 }
