--- network/src/main/java/com/sourabh/network/constants/NetworkConstants.kt
@@ -1,8 +1,0 @@
-package com.sourabh.network.constants
-
-object NetworkConstants{
-    val TMDB_API_KEY = ""
-    val TMDB_BASE_URL = "https://api.themoviedb.org/3/"
-    //Example Api Request
-//    https://api.themoviedb.org/3/movie/550?api_key=
-}
