--- network/src/main/java/com/sourabh/network/models/MoviesList.kt
@@ -1,15 +1,0 @@
-package com.sourabh.network.models
-
-
-import com.google.gson.annotations.SerializedName
-
-data class MoviesList(
-    @SerializedName("page")
-    val page: Int,
-    @SerializedName("results")
-    val results: List<Result>,
-    @SerializedName("total_pages")
-    val totalPages: Int,
-    @SerializedName("total_results")
-    val totalResults: Int
-)
