+++ repository/src/androidTest/java/com/example/repository/ExampleInstrumentedTest.java
@@ -1,8 +1,8 @@
-package com.sourabh.network;
+package com.example.repository;
 
 import android.content.Context;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.InstrumentationRegistry;
+import androidx.test.runner.AndroidJUnit4;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -21,6 +21,6 @@ public class ExampleInstrumentedTest {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getTargetContext();
 
-        assertEquals("com.sourabh.network.test", appContext.getPackageName());
+        assertEquals("com.example.repository.test", appContext.getPackageName());
     }
 }
