+++ repository/src/test/java/com/example/repository/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.sourabh.network;
+package com.example.repository;
 
 import org.junit.Test;
