+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':network'
+include ':app', ':repository'
